Remove unused `@siteimprove/alfa-json-ld` APIs
