+++ packages/alfa-json-ld/package.json
@@ -18,9 +18,7 @@
     "src/**/*.d.ts"
   ],
   "dependencies": {
-    "@siteimprove/alfa-option": "workspace:^0.17.0",
-    "@siteimprove/alfa-result": "workspace:^0.17.0",
-    "@siteimprove/alfa-url": "workspace:^0.17.0"
+    "@siteimprove/alfa-json": "workspace:^0.17.0"
   },
   "devDependencies": {
     "@siteimprove/alfa-test": "workspace:^0.17.0"
