+++ packages/alfa-json-ld/src/index.ts
@@ -1,3 +1,1 @@
-export * from "./expand";
-export * from "./guards";
-export * from "./types";
+export * from "./json-ld";
